make "go build" work
now cd src and type go build, it works. a binary named "src" is built.
(because the subdir is named src, so the binary is named src by go).
at the moment the binary executable simply prints out command line args
and exit, because I want to see if the transfer from go main to c main works.